% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{elev_from_geop}
\alias{elev_from_geop}
\title{Calculate elevation from geopotential}
\usage{
elev_from_geop(geop, RadEarth = 6371229, gravity = 9.80665)
}
\arguments{
\item{geop}{geopotential as downloaded by \code{download_ancillary_era5} can be spatraster or vector}

\item{RadEarth}{radius of earth}

\item{gravity}{speed of gravity}
}
\value{
elevation of same class as geop
}
\description{
Calculate elevation from geopotential
}
