% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_functions.R
\name{get_parcel_var}
\alias{get_parcel_var}
\title{Get parcel variable}
\usage{
get_parcel_var(
  climdata,
  var,
  parcels,
  id = "gid",
  stat = c("mean", "min", "max")
)
}
\arguments{
\item{climdata}{\itemize{
\item named list of climate timeseriess
}}

\item{var}{\itemize{
\item name of climdata element to be got
}}

\item{parcels}{\itemize{
\item vect or sf dataframe of parcel polygons
}}

\item{id}{\itemize{
\item unique id variable in parcels
}}

\item{stat}{\itemize{
\item summary statistic (min,max or mean) to be used to return single raster layer
}}
}
\value{
sf dataframe of parcels and variable values - suitable for mapping
}
\description{
\itemize{
\item get parcel values of chosen climate variable as mean/min/max values over all timesteps.
Output suitable for easy mapping using \code{plot_parcel_var} function
}
}
