% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jasmin_functions.R
\name{get_ukcp_dtm}
\alias{get_ukcp_dtm}
\title{Get ukcp18 RCM 12km DTM from ceda archive}
\usage{
get_ukcp_dtm(aoi, ukcpdtm_file)
}
\arguments{
\item{aoi}{\itemize{
\item used to cut dtm
}}

\item{ukcpdtm_file}{\itemize{
\item full path to UKCP18RCM orography/dtm file
}}
}
\value{
spatrast object cropped to extended aoi
}
\description{
Get ukcp18 RCM 12km DTM from ceda archive
}
\examples{
\dontrun{
dtmc<-get_ukcp_dtm(aoi, ukcpdtm_file=orog_filepath)
}
}
\keyword{jasmin}
