% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_functions.R
\name{map_parcel_var}
\alias{map_parcel_var}
\title{Plot parcel variable as leaflet map}
\usage{
map_parcel_var(parcels_sf, plotvar = "tmax", idvar = "gid")
}
\arguments{
\item{parcels_sf}{\itemize{
\item vect or sf object of parcels
}}

\item{plotvar}{\itemize{
\item name of variable column to be plotted in parcels_sf
}}

\item{idvar}{\itemize{
\item id variable in parcels_sf
}}
}
\value{
leaflet map object
}
\description{
Plot parcel variable as leaflet map
}
